// src/common/enums/characteristics.enums.ts
export enum EComfort {
  ONE = "1",
  TWO = "2",
  THREE = "3",
  LUX = "Lux",
}

export enum EDestination {
  RESIDENTIAL = "Rezidentiala",
  OFFICES = "Birouri",
  COMMERCIAL = "Comerciala",
  VACATION = "De vacanta",
}

export enum ECompartmentalization {
  DETACHED = "Decomandat",
  SEMI_DETACHED = "Semidecomandat",
  UNDETACHED = "Nedecomandat",
  CIRCULAR = "Circular",
  ORDERED = "Comandat",
  DUPLEX = "Duplex",
  OPEN_SPACE = "Open space",
}
export enum EBuildingType {
  BUILDING = "Bloc",
  HOUSE = "Casa / vila",
  STUDIO_APARTMENT_BUILDING = "Bloc garsoniere",
  MIXT_BUILDING = "Bloc mixt",
}

export enum EBuildingStructure {
  BRICK = "Caramida",
  CONCRETE = "Beton",
  BCA = "BCA",
  PLATES = "Placi",
  WOOD = "Lemn",
  METAL = "Metal",
  OTHERS = "Altele",
  POROTHERM = "Porotherm",
}

export enum EBuildingSeismicRisk {
  CLASS_I = "Clasa I",
  CLASS_II = "Clasa II",
  CLASS_III = "Clasa III",
  CLASS_IV = "Clasa IV",
  URGENCY_I = "Urgenta I",
  URGENCY_II = "Urgenta II",
  URGENCY_III = "Urgenta III",
}

export enum EConstructionStage {
  FINALIZED = "Finalizata",
  IN_CONSTRUCTION = "In constructie",
  PROJECT = "Proiect",
  FINISHED = "Finisat",
  SEMI_FINISHED = "Semifinisat",
  RED = "La rosu",
  REQUIRES_RENOVATION = "Necesita renovare",
  DEMOLISH = "De demolat",
  STRUCTURE = "Structura",
}
export enum EEnergyClass {
  A = "A",
  B = "B",
  C = "C",
  D = "D",
  E = "E",
  F = "F",
  G = "G",
}